Markers: Creation and Map Population
Using the google maps api to visualize the location of addresses, I have added markers to the map. These markers have geocoded latitude and longitude. The address taken from the blazor page is persisted all the way down to the database and the table showing address is coming all the way from it. This vertical slice follows a similar structure to the last: Blazor -> Web API -> Java -> DB