Reserve attempt: Dto
I made a new dto for the reservation and defined how the gRPC protobuff should work. I tried using an int alone for the foodpost, and the http patch did not like it. I tried using an empty message for the response, and it didn't work. The user still has to be added as well as the join table. However, I was able to persist the reservation status once.