Can't reserve taken food and authorized view
First, I have thrown an exception when someone tries to reserve food that does not have the state of posted. This error is brought all the way to the blazor page.
Second, the reserve button only appears if the user is logged in. The card is also moved to be inside of the card